[IOPAY-37] Refactoring
