+++ src/tasks/__test__/payTask.test.ts
@@ -1,0 +1,9 @@
+import nodeFetch from 'node-fetch';
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any,functional/immutable-data
+(global as any).fetch = nodeFetch;
+
+afterEach(() => {
+  jest.resetAllMocks();
+  jest.restoreAllMocks();
+});
